+++ src/main/java/net/geforcemods/securitycraft/blocks/KeypadBlock.java
@@ -82,6 +82,18 @@ public class KeypadBlock extends DisguisableBlock {
 	}
 
 	@Override
+	public void onRemove(BlockState state, Level level, BlockPos pos, BlockState newState, boolean isMoving) {
+		if (!state.is(newState.getBlock())) {
+			if (state.getValue(POWERED)) {
+				level.updateNeighborsAt(pos, this);
+				BlockUtils.updateIndirectNeighbors(level, pos, this);
+			}
+
+			super.onRemove(state, level, pos, newState, isMoving);
+		}
+	}
+
+	@Override
 	public boolean isSignalSource(BlockState state) {
 		return true;
 	}